Reformatting code
